backend/requirements.txt
asgiref==3.11.0
brotlipy==0.7.0
cffi==2.0.0
coverage==7.13.0
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pycparser==2.23
python-dotenv==1.2.1
sqlparse==0.5.5